backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alex-database",
    "boto3>=1.40.29",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "database",
]

[tool.uv.sources]
alex-database = { workspace = true }
